refactor: clean up sidebar internals now that it is feature complete
- Derive all sidebar widths from two CSS variables and drop the dead
  220px base width and duplicated peek-margin rules
- Share base styles between nav items, More menu items, and the More
  trigger instead of copying them
- Report tooltip/menu open state via effects so unmounting mid-popup
  can no longer leave a hover peek stuck open
- Replace the div-based resize handle with a native <button>
- Split sidebar-misc.tsx into sidebar-header/footer/group files and
  dedupe the group's accordion and plain branches
- Simplify the More context to a plain boolean and remove small DOM
  noise (data-active="false", aria-disabled="false")
