+++ packages/raystack/components/sidebar/sidebar-more.tsx
@@ -2,12 +2,12 @@
 
 import { DotsHorizontalIcon } from '@radix-ui/react-icons';
 import { cx } from 'class-variance-authority';
-import { ReactNode, useContext } from 'react';
+import { ReactNode, useContext, useEffect, useState } from 'react';
 import { Menu } from '../menu';
 import { Tooltip } from '../tooltip';
 import styles from './sidebar.module.css';
 import { SidebarLeadingVisual } from './sidebar-leading-visual';
-import { SidebarMoreProvider } from './sidebar-more-context';
+import { SidebarMoreContext } from './sidebar-more-context';
 import { SidebarPopupContext, useSidebarSafe } from './sidebar-root';
 
 export interface SidebarMoreProps {
@@ -30,7 +30,20 @@ export function SidebarMore({
 }: SidebarMoreProps) {
   const { isCollapsed, position, hideCollapsedItemTooltip } = useSidebarSafe();
   const onPopupOpenChange = useContext(SidebarPopupContext);
+  const [menuOpen, setMenuOpen] = useState(false);
+
+  // The menu portals outside the sidebar; report its open state so moving
+  // the pointer into it doesn't collapse a hover peek. An effect (not the
+  // onOpenChange handler) so unmounting mid-menu still decrements the
+  // sidebar's open-popup count.
+  useEffect(() => {
+    if (!menuOpen) return;
+    onPopupOpenChange(true);
+    return () => onPopupOpenChange(false);
+  }, [menuOpen, onPopupOpenChange]);
+
   if (!children) return null;
+
   const triggerIcon = leadingIcon ?? (
     <DotsHorizontalIcon width={16} height={16} />
   );
@@ -57,9 +70,7 @@ export function SidebarMore({
   );
 
   return (
-    // The menu portals outside the sidebar; report its open state so
-    // moving the pointer into it doesn't collapse a hover peek.
-    <Menu onOpenChange={open => onPopupOpenChange(open)}>
+    <Menu onOpenChange={setMenuOpen}>
       {isCollapsed && !hideCollapsedItemTooltip ? (
         <Tooltip>
           <Tooltip.Trigger render={<Menu.Trigger render={triggerContent} />} />
@@ -77,9 +88,7 @@ export function SidebarMore({
         className={classNames?.menuContent}
         side={position === 'left' ? 'right' : 'left'}
       >
-        <SidebarMoreProvider value={{ isInsideSidebarMore: true }}>
-          {children}
-        </SidebarMoreProvider>
+        <SidebarMoreContext value={true}>{children}</SidebarMoreContext>
       </Menu.Content>
     </Menu>
   );
