+++ packages/raystack/components/sidebar/sidebar.module.css
@@ -1,6 +1,10 @@
 .root {
+  /* Single source for the open/collapsed widths; the peek margins below
+     derive from these, so they can never drift out of sync. */
+  --sidebar-open-width: 240px;
+  --sidebar-collapsed-width: var(--rs-space-12);
+
   display: flex;
-  width: 220px;
   height: 100%;
   padding: var(--rs-space-5) 0;
   flex-direction: column;
@@ -38,20 +42,21 @@
 }
 
 .root[data-open] {
-  width: 240px;
+  width: var(--sidebar-open-width);
 }
 
 .root[data-closed] {
-  width: var(--rs-space-12);
+  width: var(--sidebar-collapsed-width);
 }
 
 /* "hidden" collapse: no icon rail, just a thin strip so the resize handle
-   stays reachable to reopen. No strip at all when collapse is disabled —
-   there's nothing to reopen from. */
-.root[data-closed][data-collapse-mode="hidden"] {
-  width: var(--rs-space-2);
+   stays reachable to reopen. Set on the mode (not on data-closed) so the
+   peek margin below sees the right collapsed width mid-peek too. */
+.root[data-collapse-mode="hidden"] {
+  --sidebar-collapsed-width: var(--rs-space-2);
 }
 
+/* No strip at all when collapse is disabled — there's nothing to reopen from. */
 .root[data-closed][data-collapse-mode="hidden"][data-collapse-disabled] {
   width: 0;
 }
@@ -67,28 +72,22 @@
 /* A peek reveals the sidebar as an overlay: the aside stays in the layout
    but a negative margin cancels everything beyond its collapsed width, so
    it paints over the adjacent content instead of pushing it. No fixed
-   positioning, so it stays anchored inside any container. The margin must
-   match the collapsed width of the active collapse mode. */
+   positioning, so it stays anchored inside any container. */
 .root[data-peeking] {
-  width: 240px;
+  width: var(--sidebar-open-width);
   z-index: var(--rs-z-index-portal);
   box-shadow: var(--rs-shadow-lifted);
 }
 
 .root[data-peeking][data-position="left"] {
-  margin-right: calc(var(--rs-space-12) - 240px);
+  margin-right: calc(
+    var(--sidebar-collapsed-width) -
+    var(--sidebar-open-width)
+  );
 }
 
 .root[data-peeking][data-position="right"] {
-  margin-left: calc(var(--rs-space-12) - 240px);
-}
-
-.root[data-peeking][data-collapse-mode="hidden"][data-position="left"] {
-  margin-right: calc(var(--rs-space-2) - 240px);
-}
-
-.root[data-peeking][data-collapse-mode="hidden"][data-position="right"] {
-  margin-left: calc(var(--rs-space-2) - 240px);
+  margin-left: calc(var(--sidebar-collapsed-width) - var(--sidebar-open-width));
 }
 
 .header {
@@ -108,9 +107,9 @@
 }
 
 .root [data-collapse-hidden] {
-  /* Cap width so the collapse animates to 0; 240px = the sidebar's open width,
-     so wider consumer content is already clipped by the sidebar itself. */
-  max-width: 240px;
+  /* Cap width so the collapse animates to 0; anything wider than the open
+     sidebar is already clipped by the sidebar itself. */
+  max-width: var(--sidebar-open-width);
 }
 
 .root[data-closed] [data-collapse-hidden] {
@@ -127,7 +126,8 @@
   gap: var(--rs-space-2);
 }
 
-.nav-item {
+.nav-item,
+.more-menu-item {
   display: flex;
   padding: var(--rs-space-3);
   align-items: center;
@@ -141,7 +141,9 @@
   box-sizing: border-box;
 }
 
-.more-trigger {
+/* Resets for elements that can render as a <button> instead of an <a>. */
+.more-trigger,
+.more-menu-item {
   width: 100%;
   border: 0;
   background: transparent;
@@ -171,12 +173,6 @@
 }
 
 .nav-item[data-active="true"] {
-  display: flex;
-  padding: var(--rs-space-3);
-  align-items: center;
-  gap: var(--rs-space-3);
-  width: 100%;
-  border-radius: var(--rs-radius-2);
   background: var(--rs-color-background-neutral-secondary);
   color: var(--rs-color-foreground-base-primary);
 }
@@ -184,24 +180,6 @@
 .nav-item[data-disabled="true"] {
   opacity: 0.5;
   pointer-events: none;
-}
-
-.more-menu-item {
-  display: flex;
-  padding: var(--rs-space-3);
-  align-items: center;
-  gap: var(--rs-space-3);
-  min-height: var(--rs-space-9);
-  align-self: stretch;
-  border-radius: var(--rs-radius-2);
-  color: var(--rs-color-foreground-base-secondary);
-  cursor: pointer;
-  text-decoration: none;
-  box-sizing: border-box;
-  width: 100%;
-  border: 0;
-  background: transparent;
-  text-align: left;
 }
 
 /* data-highlighted is the menu's arrow-key focus; keep it identical to hover
@@ -272,33 +250,36 @@
   visibility: hidden;
 }
 
-.resizeHandle {
+.resize-handle {
   position: absolute;
   top: 0;
   right: 0;
   width: var(--rs-space-2);
   height: 100%;
+  padding: 0;
+  border: 0;
+  background: transparent;
   cursor: pointer;
   transition: background-color var(--rs-duration-normal) var(--rs-ease-out);
   transform: translateX(50%);
 }
 
-.root[data-collapse-disabled] .resizeHandle {
+.root[data-collapse-disabled] .resize-handle {
   display: none;
 }
 
-.resizeHandle:hover {
+.resize-handle:hover {
   background-color: var(--rs-color-border-base-primary);
 }
 
 /* Inset ring: the handle straddles the sidebar edge (translateX(50%)), so an
    outward ring would sit partly outside the sidebar and risk clipping. */
-.resizeHandle:focus-visible {
+.resize-handle:focus-visible {
   outline: var(--rs-focus-ring);
   outline-offset: var(--rs-focus-ring-offset-inset);
 }
 
-.root[data-position="right"] .resizeHandle {
+.root[data-position="right"] .resize-handle {
   left: 0;
   right: auto;
   transform: translateX(-50%);
