+++ packages/raystack/components/sidebar/sidebar.tsx
@@ -1,10 +1,8 @@
+import { SidebarFooter } from './sidebar-footer';
+import { SidebarNavigationGroup } from './sidebar-group';
+import { SidebarHeader } from './sidebar-header';
 import { SidebarItem } from './sidebar-item';
 import { SidebarMain } from './sidebar-main';
-import {
-  SidebarFooter,
-  SidebarHeader,
-  SidebarNavigationGroup
-} from './sidebar-misc';
 import { SidebarMore } from './sidebar-more';
 import { SidebarRoot } from './sidebar-root';
 import { SidebarTrigger } from './sidebar-trigger';
